rename: public 폴더로 이동
