+++ public/css/blocks/event-banner.css
