+++ public/css/blocks/main-ad-banner.css
