+++ public/css/components/company-bar.css
