+++ public/css/components/promotion-banner.css
