+++ public/css/components/webtoon-card--big.css
