+++ public/css/components/webtoon-card--long.css
