+++ public/css/components/webtoon-card.css
