+++ public/js/blocks/expected-top.js
@@ -1,6 +1,6 @@
 import { frameHeader } from "../components/frame-header.js";
 import { webtoonBigCard } from "../components/webtoon-big-card.js";
-import { webtoons } from "../webtoons.js";
+import { webtoons } from "../data/webtoons.js";
 
 const expectedWebtoon = webtoons.expectedTop;
