+++ public/js/components/promotion-banner.js
