+++ public/js/components/sub-ad-banner.js
