+++ public/js/components/webtoon-card.js
