+++ public/js/components/webtoon-long-card.js
