+++ public/js/index.js
@@ -1,6 +1,6 @@
-import { initEventListener } from "./event-listener.js";
+import { initEventListener } from "./event/event-listener.js";
 import { dowTopDefault } from "./blocks/dow-top.js";
-import { renderingDowNavDefault } from "./rendering-nav.js";
+import { renderingDowNavDefault } from "./rendering/nav.js";
 
 const renderingDefault = () => {
   renderingDowNavDefault();
